small refactore, and progress
