+++ server/controllers/createCustomer/index.js
@@ -1,3 +1,3 @@
 module.exports = {
-  createCustomerWithLocation: require('./createCustomerWithLocation')
+  createCustomerWithLocation: require('./createLocationWithCustomer')
 }
